Basic Example Program using Fraction Class
Basic Program that shows a few of the key features of the Fraction Class, includes Visual Studio 2015 Community Solution File.
BARFraction.cs and Program.cs are included in the zip file for those who use another text editor. Additionally, the exe file is located bin/debug folder. I will be working on a windows form and asp.net application for example programs as well. They will be updated when completed.